[Refactor] network 파일 주석 제거
