+++ mohaemukzip/Service/DTOs/Auth/LoginRequestDTO.swift
@@ -1,10 +1,3 @@
-//
-//  LoginRequestDTO.swift
-//  mohaemukzip
-//
-//  Created by 이서현 on 1/31/26.
-//
-
 
 import Foundation
