+++ mohaemukzip/Service/DTOs/Home/HomeStatDTO.swift
@@ -1,9 +1,3 @@
-//
-//  HomeStatDTO.swift
-//  mohaemukzip
-//
-//  Created by 이서현 on 1/31/26.
-//
 
 import SwiftUI
