--- mohaemukzip/Service/Networking/Auth/SignUpResponseDTO.swift
@@ -1,24 +1,0 @@
-//
-//  SignUpResponseDTO.swift
-//  mohaemukzip
-//
-//  Created by 이서현 on 1/31/26.
-//
-
-
-import Foundation
-
-struct SignUpResponseDTO: Decodable {
-    let isSuccess: Bool
-    let code: String
-    let message: String
-    let result: SignUpResultDTO
-}
-
-struct SignUpResultDTO: Decodable {
-    let id: Int
-    let accessToken: String
-    let refreshToken: String
-    let isNewUser: Bool
-    let isInactive: Bool
-}
