+++ mohaemukzip/Service/Networking/Home/HomeAPI.swift
@@ -1,9 +1,3 @@
-//
-//  HomeAPI.swift
-//  mohaemukzip
-//
-//  Created by 이서현 on 1/31/26.
-//
 
 import Foundation
 import Moya
