+++ mohaemukzip/Service/Services/Auth/AuthService.swift
@@ -1,10 +1,3 @@
-//
-//  AuthService.swift
-//  mohaemukzip
-//
-//  Created by 고석현 on 1/31/26.
-//
-
 
 import Foundation
 import Moya
@@ -13,7 +6,7 @@ final class AuthService {
     // Singleton instance used across the app
     static let shared = AuthService()
     private let provider = NetworkManager.shared.makeProvider(for: AuthAPI.self)
-
+    
     func signup(nickname: String, loginId: String, password: String, terms: [SignUpTermDTO]) async throws -> AuthTokensModel {
         let requestDTO = SignUpRequestDTO(
             nickname: nickname,
@@ -21,7 +14,7 @@ final class AuthService {
             password: password,
             terms: terms
         )
-
+        
         return try await withCheckedThrowingContinuation { continuation in
             provider.request(.signup(requestDTO)) { result in
                 switch result {
@@ -32,7 +25,7 @@ final class AuthService {
                     } catch {
                         continuation.resume(throwing: error)
                     }
-
+                    
                 case .failure(let error):
                     continuation.resume(throwing: error)
                 }
@@ -42,7 +35,7 @@ final class AuthService {
     
     func login(loginId: String, password: String) async throws -> AuthTokensModel {
         let requestDTO = LoginRequestDTO(loginId: loginId, password: password)
-
+        
         return try await withCheckedThrowingContinuation { continuation in
             provider.request(.login(requestDTO)) { result in
                 switch result {
@@ -50,29 +43,29 @@ final class AuthService {
                     do {
                         let decoded = try response.map(LoginResponseDTO.self)
                         let model = decoded.result.toModel()
-
+                        
                         // 정상 동작: 로그인 성공 시 토큰을 저장하고 상위(AppState 등)에서 화면 전환을 처리한다.
                         // 아래 TEST ONLY 블록은 "access 토큰이 깨진 상태에서 401 → reissue → 재시도" 흐름이
                         // 자동으로 동작하는지 확인하기 위한 테스트 코드다.
                         // - 평소엔 주석 상태로 두고,
                         // - 테스트할 때만 주석을 풀어 사용한 뒤 반드시 제거한다.
                         /*
-                        // ================================
-                        // TEST ONLY: accessToken 강제 오염
-                        // 자동 reissue 동작 확인용
-                        // ================================
-                        let corruptedAccess = String(model.accessToken.dropLast(8)) + "TESTTEST"
-
-                        // UserDefaults/Keychain 저장소에 오염된 access를 덮어쓴다.
-                        TokenStore.saveTokens(access: corruptedAccess, refresh: model.refreshToken)
-
-                        // 런타임에서 사용하는 Config도 함께 갱신한다.
-                        Config.accessTK = corruptedAccess
-                        Config.refreshTK = model.refreshToken
-
-                        print("[TEST] access token intentionally corrupted")
-                        */
-
+                         // ================================
+                         // TEST ONLY: accessToken 강제 오염
+                         // 자동 reissue 동작 확인용
+                         // ================================
+                         let corruptedAccess = String(model.accessToken.dropLast(8)) + "TESTTEST"
+                         
+                         // UserDefaults/Keychain 저장소에 오염된 access를 덮어쓴다.
+                         TokenStore.saveTokens(access: corruptedAccess, refresh: model.refreshToken)
+                         
+                         // 런타임에서 사용하는 Config도 함께 갱신한다.
+                         Config.accessTK = corruptedAccess
+                         Config.refreshTK = model.refreshToken
+                         
+                         print("[TEST] access token intentionally corrupted")
+                         */
+                        
                         continuation.resume(returning: model)
                     } catch {
                         continuation.resume(throwing: error)
@@ -83,12 +76,12 @@ final class AuthService {
             }
         }
     }
-
+    
     // MARK: - Check LoginId (Duplicate Check)
-
+    
     func checkLoginId(loginId: String) async throws -> CheckLoginIdModel {
         let requestDTO = CheckLoginIdRequestDTO(loginId: loginId)
-
+        
         return try await withCheckedThrowingContinuation { continuation in
             provider.request(.checkLoginId(requestDTO)) { result in
                 switch result {
@@ -105,13 +98,13 @@ final class AuthService {
             }
         }
     }
-
+    
     func checkDuplicateId(loginId: String) async throws -> Bool {
         try await checkLoginId(loginId: loginId).available
     }
-
+    
     // MARK: - Token Reissue
-
+    
     /// AccessToken 만료 등으로 401 발생 시 사용
     /// - Returns: 새로 발급된 (accessToken, refreshToken)
     func reissue() async throws -> (accessToken: String, refreshToken: String) {
@@ -125,16 +118,16 @@ final class AuthService {
                     } catch {
                         continuation.resume(throwing: error)
                     }
-
+                    
                 case .failure(let error):
                     continuation.resume(throwing: error)
                 }
             }
         }
     }
-
+    
     // MARK: - Logout
-
+    
     func logout() async throws {
         return try await withCheckedThrowingContinuation { continuation in
             provider.request(.logout) { result in
@@ -147,9 +140,9 @@ final class AuthService {
             }
         }
     }
-
+    
     // MARK: - Withdrawal
-
+    
     func withdrawal() async throws {
         return try await withCheckedThrowingContinuation { continuation in
             provider.request(.withdrawal) { result in
