+++ mohaemukzip/Service/Services/Home/HomeService.swift
@@ -1,9 +1,3 @@
-//
-//  HomeService.swift
-//  mohaemukzip
-//
-//  Created by 이서현 on 1/31/26.
-//
 
 import Foundation
 import Moya
